Split io.py into bids.py, caps.py and filemanip.py
This commit removes io.py file in order to avoir ambiguous situtation e.g. in Issue #37. Functions in io.py have moved:
- Move check_{bids|caps}_folder, {bids|caps}_type_to_path, {bids|caps}_type_to_description, check_input_{bids|caps}_files functions to the new {bids|caps}.py file
- Move fix_join to the new nipype.py file (the function is needed for several build_output_node methods)
- Move zip_nii and unzip_nii to the new filemanip.py file (inspiration from nipype)
- Move save_participants_sessions, extract_crash_files_from_log_file, read_participant_tsv, get_subject_session_list, reorder_bids_or_caps_files, get_subject_id and extract_image_ids to filemanip.py
