+++ clinica/pipelines/engine.py
@@ -77,7 +77,9 @@ class Pipeline(Workflow):
         import inspect
         import os
         from colorama import Fore
-        from clinica.utils.io import check_bids_folder, check_caps_folder, get_subject_session_list
+        from clinica.utils.caps import check_caps_folder
+        from clinica.utils.bids import check_bids_folder
+        from clinica.utils.filemanip import get_subject_session_list
         from clinica.utils.exceptions import ClinicaException
 
         self._is_built = False
