+++ clinica/pipelines/fmri_preprocessing/fmri_preprocessing_pipeline.py
@@ -317,7 +317,8 @@ class fMRIPreprocessing(cpe.Pipeline):
         import nipype.interfaces.utility as nutil
         import nipype.interfaces.spm as spm
         import nipype.pipeline.engine as npe
-        from clinica.utils.io import zip_nii, unzip_nii
+        from clinica.utils.filemanip import unzip_nii
+        from clinica.utils.filemanip import zip_nii
 
         # Zipping
         # =======
