+++ clinica/pipelines/t1_volume_new_template/t1_volume_new_template_pipeline.py
@@ -138,7 +138,7 @@ class T1VolumeNewTemplate(cpe.Pipeline):
         import nipype.pipeline.engine as npe
         import nipype.interfaces.io as nio
         from ..t1_volume_tissue_segmentation import t1_volume_tissue_segmentation_utils as seg_utils
-        from clinica.utils.io import zip_nii
+        from clinica.utils.filemanip import zip_nii
         import os.path as op
         import re
 
@@ -311,7 +311,7 @@ class T1VolumeNewTemplate(cpe.Pipeline):
         from ..t1_volume_create_dartel import t1_volume_create_dartel_utils as dartel_utils
         from ..t1_volume_dartel2mni import t1_volume_dartel2mni_utils as dartel2mni_utils
         from ..t1_volume_parcellation import t1_volume_parcellation_utils as parcellation_utils
-        from clinica.utils.io import unzip_nii
+        from clinica.utils.filemanip import unzip_nii
         from clinica.utils.spm import get_tpm
 
         # Get Tissue Probability Map from SPM
